Avoid naming conflicts in raw API
When the plural and the singular raw name was the same,
the deduper kicked in and added a `0`, e.g.
```swift
func sqlite3_talents0_fetch()
```

This happened because the `talents` was plural already
(and hence ended up with the same name).
This only happens w/ `singularizeRecordNames` set to
false (the default).

Fixes issue #12.