Open TX w/ immediate by default
The previous default was `deferred` which can
upgrade transactions, but has the problem that
it'll raise SQLITE_BUSY immediatly on upgrades,
i.e. the busy handler is not being used.

As per: https://kerkour.com/sqlite-for-servers

This is a relevant change to behaviour. It
also means, that one should more proactively
use `readTransaction` to avoid over-locking.