Fix issue 27: .sql/.db could not be marked as a resource in Xcode
This bug was introduced due to a fix in apparently
Xcode 15.4.
Previous versions of Xcode would not copy the resource
into the target, hence Enlighter did that part.
This seems to be fixed in Xcode now, so it ends up with duplicate
resource files.
Only fixed that for Xcode 16.

Note: If the resource is required in an existing project, this will
now need to mark the resource as a copy-resource again in Xcode 16+.